Move the first test to the common trait
